refactor(core): move processing modules to core package
