+++ src/core/__init__.py
@@ -1,0 +1,7 @@
+"""Core modules for the AIOCR project."""
+
+from .db_manager import DBManager
+from .template_manager import TemplateManager
+from .ocr_agent import OcrAgent
+
+__all__ = ["DBManager", "TemplateManager", "OcrAgent"]
