+++ src/core/ocr_processor.py
