+++ tests/test_preprocess.py
@@ -8,7 +8,7 @@ import os
 import sys
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '../src')))
 
-from app import preprocess
+from core import preprocess
 
 @pytest.fixture
 def sample_image():
